+++ src/main/messages.module/auth-error.ts
@@ -1,5 +1,4 @@
 import { ServerMessage } from './server-message'
-import { AuthErrorEnum } from './auth-error-enum'
 export interface AuthError extends ServerMessage {
 
 }