+++ src/main/messages.module/ownership-error.ts
@@ -1,0 +1,7 @@
+import { GeneralError } from './general-error'
+
+export const OwnershipError: GeneralError = {
+    message: 'Você não tem direitos sobre esse item',
+    name: 'OwnershipError',
+    status: 403
+}