+++ src/main/models.module/base-model.ts
@@ -1,6 +1,11 @@
+import { ObjectIdClient } from './object-id-client'
+import { ObjectIdServer } from './object-id-server'
+
+
 export interface BaseModel {
     /**
-     * Identifier internally used by mongoose on backend
+     * Identifier internally used by mongoose on backend.
+     * It is a string on frontend and an ObjectId object on backend
      */
-    _id: string
+    _id: ObjectIdClient | ObjectIdServer
 }