+++ src/main/models.module/bau-card.ts
@@ -1,10 +1,11 @@
 import { BaseModel } from './base-model'
-import { User } from './user'
+import { ObjectId } from './object-id'
 import { BauTag } from './bau-tag'
+import { User } from './user'
 
-export interface BauCard extends BaseModel{
+export interface BauCard extends BaseModel {
     question: string
     answer: string
-    tags: BauTag[]
-    createdBy: User
+    tags: (ObjectId | BauTag)[]
+    createdBy: (ObjectId | User)
 }