+++ src/main/models.module/bau-tag.ts
@@ -1,7 +1,8 @@
 import { BaseModel } from './base-model'
 import { User } from './user'
+import { ObjectId } from './object-id'
 
 export interface BauTag extends BaseModel {
     text: string
-    createdBy: User
+    createdBy: (ObjectId | User)
 }