+++ src/main/models.module/minecraft-step.ts
@@ -1,8 +1,10 @@
 import { BaseModel } from './base-model'
+import { ObjectId } from './object-id'
+import { User } from './user'
 export interface MinecraftStep extends BaseModel {
     x: number
     y: number
     z: number
-    user: string
+    user: (ObjectId|User)
     date: Date
 }