+++ src/main/models.module/object-id-server.ts
@@ -1,0 +1,3 @@
+import mongoose = require('mongoose')
+
+export type ObjectIdServer = mongoose.Types.ObjectId