+++ src/main/models.module/object-id.ts
@@ -1,0 +1,4 @@
+import { ObjectIdServer } from './object-id-server'
+import { ObjectIdClient } from './object-id-client'
+
+export type ObjectId = ObjectIdServer | ObjectIdClient